Более корректное ведение списка последних пользователей

+++ modules/kto_zloy.py
@@ -1,7 +1,5 @@
 import random
-import logging
 
-from peewee import SQL, fn
 from telegram.ext import CommandHandler, Filters, MessageHandler
 
 from model import LastUsers, User
@@ -9,10 +7,12 @@ from utils import get_username_or_name
 
 
 class KtoZloy:
+    _LAST_USERS_MAX_SIZE = 10
+
     def __init__(self, chat_id, admin_id):
         self._chat_id = chat_id
         self._admin_id = admin_id
-    
+
     def add_handlers(self, add_handler):
         add_handler(CommandHandler('ktozloy', self._kto_zloy))
         add_handler(MessageHandler(Filters.all, self._update_last_users))
@@ -25,18 +25,20 @@ class KtoZloy:
         user, _ = User.get_or_create(user_id=message.from_user.id, defaults={
             'username': get_username_or_name(message.from_user)})
 
+        LastUsers.delete().where(LastUsers.user == user).execute()
+
         LastUsers.create(user=user)
 
-        border_id = self._get_last_users()[-1].last_id
-        
-        logging.warning("border_id = " + str(border_id))
-        
-        LastUsers.delete().where(
-            LastUsers.id < border_id).execute()
-    
+        last_ids = [row.id for row in LastUsers.select(LastUsers.id).order_by(LastUsers.id.desc())]
+
+        if len(last_ids) > self._LAST_USERS_MAX_SIZE:
+            border_id = last_ids[self._LAST_USERS_MAX_SIZE - 1]
+            LastUsers.delete().where(
+                LastUsers.id < border_id).execute()
+
     def _kto_zloy(self, bot, update):
         message = update.message
-        
+
         chat_id = message.chat_id
 
         if chat_id != self._chat_id:
@@ -46,20 +48,10 @@ class KtoZloy:
             bot.sendMessage(chat_id=chat_id, text='я злой ¯\_(ツ)_/¯')
             return
 
-        last_usernames = [row.username for row in self._get_last_users()]
-        
+        last_usernames = [row.username for row in User.select(User.username).where(User.username != "")]
+
         random_user = random.choice(last_usernames)
 
         if random_user == get_username_or_name(message.from_user):
             random_user = 'ты'
         bot.sendMessage(chat_id=chat_id, text='%s злой' % random_user)
-
-    @staticmethod
-    def _get_last_users():
-        return list(
-            User.select(User.username, fn.MAX(LastUsers.id).alias("last_id"))
-                .join(LastUsers)
-                .where(User.username != "")
-                .group_by(User.id, User.username)
-                .order_by(SQL("last_id").desc())
-                .limit(10))
